+++ main/CommonFunctions.f90
@@ -34,11 +34,11 @@ contains
     end function riMinusHalf
 
     ! Omega CommonFunctions.
-    ! Returns the complex(16) result of the Omega function in the point (r[i]).
+    ! Returns the complex(8) result of the Omega function in the point (r[i]).
     function Omega1(i) result(retval)
         implicit none
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = C1 * e ** (-ri(i) ** 2 / AC ** 2)
     end function Omega1
@@ -46,7 +46,7 @@ contains
     function Omega1Conjg(i) result(retval)
         implicit none
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = conjg(Omega1(i))
     end function Omega1Conjg
@@ -54,7 +54,7 @@ contains
     function Omega2(i) result(retval)
         implicit none
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = C2 * e ** (-ri(i) ** 2 / AC ** 2)
     end function Omega2
@@ -62,7 +62,7 @@ contains
     function Omega2Conjg(i) result(retval)
         implicit none
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = conjg(Omega2(i))
     end function Omega2Conjg
