+++ main/EquationSystemUtils.f90
@@ -10,9 +10,9 @@ module EquationSystemUtils
 contains
     function Ro23Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = IMG_UNIT * Omega1(i) * Ro12PointConjg(roMeshMatr, i) / DeltaStroke21 &
             - IMG_UNIT * Omega1(i) * (Ro33Point(roMeshMatr, i) - Ro22Point(roMeshMatr, i)) / DeltaStroke21
@@ -20,9 +20,9 @@ contains
 
     function Ro13Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = IMG_UNIT * Omega2(i) * Ro12Point(roMeshMatr, i) / DeltaStroke11 &
             - IMG_UNIT * Omega1(i) * (Ro33Point(roMeshMatr, i) - Ro11Point(roMeshMatr, i)) / DeltaStroke11
@@ -30,81 +30,81 @@ contains
 
     function Ro11Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_11, i)
     end function Ro11Point
 
     function Ro22Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_22, i)
     end function Ro22Point
 
     function Ro33Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_33, i)
     end function Ro33Point
 
     function Ro12Point(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_12, i)
     end function Ro12Point
 
     function Ro11PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_11_CONJG, i)
     end function Ro11PointConjg
 
     function Ro22PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_22_CONJG, i)
     end function Ro22PointConjg
 
     function Ro33PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_33_CONJG, i)
     end function Ro33PointConjg
 
     function Ro12PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = roMeshMatr(RO_12_CONJG, i)
     end function Ro12PointConjg
 
     function Ro23PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = -IMG_UNIT * conjg(Omega1(i)) * Ro12Point(roMeshMatr, i) / DeltaStroke22 &
             + IMG_UNIT * conjg(Omega1(i)) * (Ro33PointConjg(roMeshMatr, i) - Ro22PointConjg(roMeshMatr, i)) / DeltaStroke22
@@ -112,9 +112,9 @@ contains
 
     function Ro13PointConjg(roMeshMatr, i) result(retval)
         implicit none
-        complex(16), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
+        complex(8), dimension(1:SYSTEM_VAR_NUM, 0:N), intent(in) :: roMeshMatr
         integer :: i
-        complex(16) :: retval
+        complex(8) :: retval
         
         retval = -IMG_UNIT * conjg(Omega2(i)) * Ro12PointConjg(roMeshMatr, i) / DeltaStroke12 &
             + IMG_UNIT * conjg(Omega1(i)) * (Ro33PointConjg(roMeshMatr, i) - Ro11PointConjg(roMeshMatr, i)) / DeltaStroke12
